Moving debug actions to their own file
